Implemented a working version of rhosigchi. It needs testing, and I need to add the option to use the original rhosigchi with the first generation variance estimation inside itself.
I also began a generate_ensemble() method. But I need to think about the structure. Every ensemble member should probably be treated as its own instance of the pyma class. However, the fit() method will require the output (variance matrix) from the generate_ensemble() class. So maybe I need to split it in three? A class for unfolding+firstgen, a separate class for the rho+T fit, and a class for the ensemble generation?