The MC module is now working, at least in principle. It perturbs spectra and saves to file, and also makes a first generation variance matrix and saves that.
There are a number of things I need to check now: Does rhosigchi work correctly? Does the first-generation method work correctly? Does the unfolding (sans compton subtraction) work correctly?

Also, I have to think about the structure of the program. It may be a little silly to not have everything in the same class, with maybe pymama as a subclass. That way we could have fname_resp_mat, var_firstgen and other such global variables in the class, to be used by e.g. the fit() method. But maybe it breaks some other logic. Have to think about it more.