Refactor code
The Matrix and Vector class are moved from library.py to matrix.py,
and large portions of Matrix is rewritten to be more monolithic. Some functions
in library.py are renamed to be more meaningful.

The old tests have been moved to the temporary oldtests directory while they
are being rewritten to use pytest. A such example is the test_matrix.py test
which tests some basic properties of the Matrix class.

The readme as been updated and Travis CI is attempted to be used