This is a working version implementing rhosig.py as the fitting module. It wraps rhosig.py into the fit_rho_T() function to make it easier to use in the current workflow.
It should be cleaned up -- notably, since rhosig.py uses middle-bin energy arrays, I have just implemented some back-and-forth transformations in a rather hacky way. On the todo-list is to change everything to lower-bin-edge for consistency. Also, I don't think I'm handling the cutting of the diagonal in a proper way yet.

I have updated the notebook example so that it demonstrates the current working version. It's gratifying to see that it seems to give something sensible (at least to my eyes) for the Dy164 fit!