Fix some bugs
I got confused by the naming of E0 and E1, so they have been changed
to Eg and Ex for gamma energy and excitation energy. A figure was made
to illustrate how these relate to the matrix and the plot for future
readers.

Attempted to unify the call signature of the different methods, but
a lot of work remain.