Complete Ensemble class
The MatrixAnalysis class is redundant and is pending removal in
a later change once all necessary details have been taken care of.

The ErrorPropagation class is rewritten into Ensemble. The design of
Ensemble is for the user to specify the unfolding and first generation methods
using custom Unfolder and FirstGeneration instances which are given to
Ensemble as

```python
    unfolder = Unfolder()
    unfolder.cut_diagonal(min, max)
    firstgen = FirstGeneration()
    firstgen.dEg = 800
    ensemble = Ensemble(raw)
    ensemble.unfolder = unfolder
    ensemble.firstgen = firstgen
    ensemble.generate(number=10)
```

This makes it easy and flexible to micromanage the ensemble generation.

Work has begun on the FirstGeneration class, but is not near ready.
In the meantime, use a lambda to enclose a call to first_generation_method.

Minor fixes as been done all around the code:
 * Matrix now accepts [] for getting and setting Matrix.values
 * Matrix.state can be set by a string which is then converted into a
   MatrixState
 * Adjusted the call signature of some functions for consistency