Fix a few bugs. Clean up code
The Enum class in Python turns out to be a horrendous class
when used across modules. A semi-solution is used whereby comparison
is made using MatrixState.value and type is done directly via type().
Also moved to new file in an attempt to minimize the behavior.

Renamed a few functions to be more practical.

Unfolder now checks if Matrix.mask exists, and uses that if no
other mask is given. Matrix.cut resizes its mask.

Wrote docstrings for Ensemble.

Cleaned up notebook to use "modern OMpy".

fillnegative() and remove_negative() are called in Unfolder.unfold()
and first_generation_method(), as this seems to be what Jørgen did.
Find a better way to do it.