Continue work on rewriting the first generation method
All of the initial setup for the first generation method is completed. What
remains is the actual loop itself. This will be written as the unfolding loop,
where the looping is separated from the first generation logic in a separate
step.

During the rewriting it has been discovered that a lot of the code was either
unnecessary, such as reshaping the input matrix, or redundant, such as
constructing the normalization matrix different from the initial weights
matrix.

Some of the original names in MAMA was typical Fortran abominations, and have
been remained to something more intuitive.

As some operations are repeated throughout the code, such as summing along
one axis or cutting the diagonal, these has been abstracted away into
methods in Matrix, so as the intent is more clear to the reader instead of
just the bare bones calculation.

The rebin function was slightly touched upon, but more work remains to make it
more readable.


Former-commit-id: 042e787ab73479f15262d285f260eede88e74694