Make Matrix.plot() use midbin and add Matrix.rebin()
The Matrix.plot() now plots using the midpoint of a bin. The
solution is rather hacky, but as of now necessary since
pcolormesh() sometimes uses the midbin, sometimes uses another edge,
hence it is not possible to use a simple shift of the ticks.
It is suggested to look into the use of matplotlib.ticker to
make the ticks dynamic during zoom.

The rebin functions has been slightly updated to be more readable.
They can be rewritten in the future if the need arises.

A wrapper around rebin_2D() is added to Matrix, making the manipulation much
more high level. The method supports a 'factor' to multiply each step with,
making it easy to rebin a matrix. For now, only one axis can be
rebinned at a time, but that might change in the future.