Complete first generation method
The first generation method is believed to be complete.
There are some differences to both Jørgen's and Magne's code, as some
different decisions were made. Both had a bug where they used lower bin
to calculate the multiplicities, while it should have been mid bin.
The initial setup of H and W was erroneous as they should have been row
normalzied from the diagonal, not for the entire row.

The "slide-hack" by Magne was chosen to be disabled by default.

A worry is that no method works on idealized cases, and so it is difficult to
check whether the method actually works as intended. Further consideration is
necessary.

Some tests have finally been written using PyTest. They are working as intended,
and caught a bug in how the matrices are cut.

The Matrix now uses mid bin by default, with easy conversion using the
to_mid_bin() and to_lower_bin().

The ticks in the plotting is still not ideal. It has proven to be a far more
annoying problem than initially expected. We have to write our own Ticker
to make it work.

A new module has been written, hooks.py, which is intended for introspection
of what the code is doing. hooks.plot_hook() will attempt to plot each argument
given, for example, and is useful for inspecting how a method is manipulating
matrices. A benefit of this concept is that it forces even more standardization
across the different classes in order to be compatible with the hook system,
which again makes everything more readable.

Several small fixes and changes has been performed all around the code base,
in an attempt to standardize the calling interfaces.


Former-commit-id: 44c0418d01fa83eab34746c7f64c40458a6c983c