Begin work on Extractor
A new class, Extractor, is being written as a high level interface
to fit_rho_T and rhosig. It will take care of the necessary fitting
given an Ensemble instance.

To support the new class, the Vector class has been expanded to be more
flexible like its big brother Matrix, and a two new library functions are
written to save a Vector or Matrix to numpy binary files stored in
tar files.

Git large file storage is being tried.