Begin major rewrite of rhosig
A new branch is created in order to break rhosig before rebuilding.

Minor fixes has been done throughout the code to make the usage of the
different modules more consistent.

Saving and loading of Matrix, Vector, Ensemble and Extractor has been
greatly improved to ease the use of partial analysis.


Former-commit-id: cc117fa48d74b28d7109ee8924ce49eacc9c402b