Complete Extractor
A new class, Extractor, is implemented which decomposes matrices into
nld and gsf. It works in a similar fashion to Ensemble, with ensemble
members and book keeping. So similar, in fact, that they can be merged
into one at a later stage if that is required.

The rhosig.pyx file is rewritten to decomposition.pyx, with cleaner
syntax, multithreading and memory views. All of the functions are faster
than their equivalents. The index() method is used in Matrix for faster
indexing (~9x).

The Action class is incoporated into FirstGeneration, and Unfolder has
been greatly simplified. Setable seems to have a bug where it doesn't
correctly identify an attribute as 'modified'.

Some questions have arisen: How should the response matrix be binned
when using in the Unfolder? How should the "zeroes below the diagonal"
be specified? How extensive shall Ensemble be? How do we verify that the
minimization technique in Extractor works as intended?

Regarding the last question, the experimental matrices are very
difficult to minimize, so verification must be done on toy matrices.
This is done in a notebook using creative functions as nld and gsf. The
method indeed seems to work, but there is often a systematic deviation
from the Fact. Does this imply that there are several solutions for
diagonal function decomposition? If so, should we explore the solution
space more?

Another worry that arrises is what do we mean by "uncertainty" in nld
and gsf? Defined as the uncertainty due to uncertainty in the input
matrix (i.e, current Ensemble solution) then this is trivial, but this
doesn't actually reflect the true uncertainty. This is due to the
minimization itself. The product nld*gsf is never equal to an
experimental matrix as there are structures in the True Martix that is
not captured in the experimental data, and so the minimization can
therefore punish discreepancies between what nld*gsf shows versus what
the experimental data show. Should the minimization be restricted to
areas of good statistical data?