Complete Extractor, fix several small bugs
Extractor is completed and tested against rhosig. Everything seems to
work, but there are several oddities that have come to light when we
have thought harder about what is actually happening. Some of these are
as of yet unresolved. See GitHub issues for more indepth explanation.

The quest for uniform call signatures and class interfaces is ever
ongoing, with several small fixes written all around. Notable mentions
are the 'path' attributes which will try to load if set from init and
not complain about missing directory.