Closes #3, response function interpolation
The issue was reopened, but actually, just a wrong fwhm was chosen. The rest worked fine (one 1 call signature was changed).

However, when looking at it, I found another bug. The response around and below the backscatter peak was modeled rather bad. It was improved now. Still, for best results, I'd recommend not to interpolate the response matrix, but load it in the binning you need.

Otherwise, I also corrected the `project` method in the `Matrix` class. If one projects on the `Eg` axis, one want to have a spectrum of with the length of `Eg`, not `Ex`.


Former-commit-id: ff44cef706dcb6a9d8b0cd9163cc6ed5b1682f4f