Autodisable openMP on macOS (#51)
* Autodetecting OS

If the openMP env is not set, it will check if the OS is MacOS. If so, do not build with OpenMP.

* Fixed OpenMP issue for macOS (issue #30).
I've added some steps that will check if the libomp is installed and if so add the correct compiler flags.
It still respects the variable 'ompy_OpenMP'. The OpenMP can easily be installed with `brew install libomp`.