Important fix to unfolding. Remove self.zeroes
(Main)Problem:
The unfolder had a default attribute `self.zeroes` that
would delete (set to 0) all counts in folding with a certain mask. By
default this was set to `self.raw.diagonal_mask()`. The diagonal mask
enforces a cut based on a Ex dependend resolution. If the spectra are
calibrated incorrectly, eg. negative Ex, this will lead to unexpected
results.

Solution:
removed totally. User shall perform all necessary cuts before sending
the spectrum to unfolder.

Otherwise: Small change in chi2 and a fix in the fluctuations code.
Still, this should be checked more carefully if one likes it.