Reverted decomposition bug for bad fix 0b5e3c7
The fix was was erroneous. The issue was not due to lower-bin to mid-bin conversion. Instead, the half-bin shift should be there.

reason for the shift:
assume `Ex = 1`, `Eg = 2`, `resolution = 0.7` and a binwidth (not halved) `dE =  1`. As `Ex + resolution` **extends into** the Eg bin `Eg=2` we get following selections:
- `Eg[i_Eg] < Ex + resolution = 2 < 1 + 0.7  = False`
- `Eg[i_Eg] < Ex + resolution + dE/2 =   2 < 1 + 0.7 + 0.5  = True`
So we want the latter selection