Fixes #85 by showing always the current git sha
The only problem with the current setup could be, that the version number may not match the git commit, as the version number is retrieved from setup.py's auto generated setup_version.py file.

The setup_version.py file can also be used for trouble shooting purposes, so to see when people have last run a `build` or `install` for the package (if cython modules changed).