Fixes #107; Problem with varying sizes of ensemble matrices
After parallelization in #94, we had an issue if a matrix had negative Ex entries. They were automatically cut in the fg method. This leads to different sizes of the raw, unfolded and firstgen ensembles -- which created a mess later.

Instead, now the firstgeneration method throws an error for if the input matrix has negative excitation energies.
Additionally, there is a assert statement of the `step` function in ensemble.