Fixes #61, bug fix and change of fill_negative functions
Fixed buggy fill negatives function and replaced it by a new method,
```
Fill negative channels with positive counts from neighboring channels.

The idea is that the negative counts are somehow connected to the (γ-ray)
resolution and should thus be filled from a channel within the resolution.

This implementation loops through the closest channels with positive number
of counts to fill the channle(s) with negative counts. Note that it can
happen that some bins will remain with negative counts (if not enough bins
with possitive counts are available within the window_size) .

The routine is performed for each Ex row independently.
```

Additionally changed code of FirstGeneration and Unfoder to have
easier access if someone want to replace the fill negatives function
by a custom function