Fixes #61, bug fix and change of fill_negative functions
Fixed buggy fill negatives function and replaced it by two new methods,
either filling from a maximum closeby channel, or from the weighted maximum. where the weights are given by a gaussian function

```
Fill negative channels with positive counts from neighboring channels.

The idea is that the negative counts are somehow connected to the (γ-ray)
resolution and should thus be filled from a channel within the resolution.

This implementation loops through the closest channels with positive number
of counts to fill the channle(s) with negative counts. Note that it can
happen that some bins will remain with negative counts (if not enough bins
with possitive counts are available within the window_size) .

The routine is performed for each Ex row independently.
```