Fixes #121 by increasing unfolder iterations ++
We have see nthat the unfolding routine selected low number of iterations, which lead to non-satisfactory results. I have made seveal changed such that the selected iteration is higher, mainly:

- decreased the weight of the fluctuations
- increased the minim. number of iterations
- (also the def. number of iteratios run is not 200)

In addition we use the Chi2, not reduced chi2 in the scoring routine. The fluctuations are now cacluated based on the relative fluctuations, not absolute fluctuations, mimicing mama better.

I don't think that this is *very* satisfying, as the parameters are not well tested. But it seems to do the trick -- and the parameters have not been well studied before either.