Version pinning in Dockerfile (#124)
After a hard struggle I was able to use a custom image (not jupyter/minimal or so) with MyBinder. What a struggle, and I'm not sure it's totally worth it. Anyhow, the idea was to have a common Dockerfile for CodeOcean and MyBinder. At least to the extend possible. Might have been much easier to just have two Dockerfiles.

Challenges: MyBinder wants to enter the notebook at a certain state, and resets NB_USER. I never succeeded with the ENTRYPOINT definition, but the image works now.