Slight tweak
I've tweaked the method to ignore the `units` keyword when saving to MaMa format since the MaMa format requires keV. Also added a warning whenever a user tries to use the option together with the mama format. A small note on the format has been added to the readme file.