Refactor Response
Begin work on refactoring the massive response code. The design and
interface is significantly simplified by dividing up the class.
`ResponseData` loads the response data from file and normalizes the
counts to probabilities, as well as handling weight corrections.
`ResponseInterpolation` handles the interpolation of the discrete peaks.
Is currently only a linear interpolation, but is going to switch over to
GSF3 and logpolynomials.
The `compton.py` file contains all functions related to interpolating
the compton, by far the most complicated code. It improves on the
previous code, and MAMA, by handling the different interpolations more
carefully. It is unresolved how best to do this, but as for now a
logistic weighting is used (idea by Kvellestad). Numerical precision of
the raw count data is visible, and needs to be improved.
`Response` simply gathers the different behaviors and deals with
book keeping.

Numba is used to speed up the code, as the pure python code takes ages.
A speed up of about three orders of magnitude is seen. Moving the
compton to GPU should be even faster.

The Vector and Matrix classes are overdue for an overhaul, as their
design decisions about binning is making life miserable for more
complicated code. The indexing is to be moved into a `Index` class that
deals with all indexing and binning. Most of the scaffolding is
completed in `index.py`.