Added the 'rhosigchi' algorithm
I've re-implemented the 'rhosigchi' algorithm used to extract NLD and GSF from the first generation matrix. The time consuming iterations are done in C++.
The updates are completely backwards compatible and only adds an  additional choice.
There are still some debugging left and I still need to add propper tests of the code.