Workflow (#96)
* Debugging git workflow to replace travis

* Updates to workflow

* Debugging...

* Fixed a small bug

In the latest linux versions there seems to be an error where the array with size 1x1 is not converted into a float. We now explicitly convert to a float. The tests now passes.

* Added a macOS test as well

* Debugging

* Slight error...

* Updated workflow

* Publish workflow

Added a new workflow to publish to pypi. For now this will push to the test repository until we are ready to create official releases.

* Merging stats.cpp and erfinv.hpp

I've merged the stats.cpp and erfinv.hpp as the code could easily live in the same file.

* Modified setup file

I've modified the setup files in preparation for distribution on pypi.

* Fixing a syntax error

* Update to use a different image

pypi requires a 'manylinux' build. Hopefully this will build that.

* More adjustments to the workflow...

* More fixes...

Too bad you can't debug locally...

* Okei, this should be it!

* Additional changes

More changes in order to facilitate simple integration of cibuildwheel for building all the wheels

* A little more debugging...

* Testing the cibuildwheel solution

* Working on determining where stuff ends up...

* Update main.yml

* Added test step

* Missing &&

* Moved configuration to pyproject

I've moved the configuration parameters for cibuildwheel to the pyproject file.

* Modification to the example code

I've changed the structure of the example code to make sure that it gets copied when building the wheel

* Build and test both macOS and linux

* Updates to the build settings

* Added missing environmental variable FC

Debugging macOS build

* Added the example matrix files

* No install because of issues while installing

* Removed some compiler flags that does not work on apple silicon

* Test deployment too

* Testing with ubuntu only...

* Removed the upload job. Will add once I know that the test release works

* Added python tag

* Stupid errors...

* Added missing upload