Refactor of Response almost complete.
The hard parts of the response are refactored into Response,
ResponseData, DiscreteInterpolation and compton. It is much, much
simpler than the old design and favors a new format with much faster io,
more accurate interpolation, somewhat faster interpolation.

The new Index class is nearing primetime, but in the meantime rebinning
is broken for 2D.

Features missing:
- Refactor rebinning
- Expand rebinning (low priority)
- Refactor Matrix
- Refactor AbstractArray operations
- Complete Response (highest)