Continue work on the Response, rewrite Unfolder
The Response is fast and usable, but has a few wrinkles. Specializing to
an Index beyond the span of the Response gives a shift in the response
matrix, probably a bug in how 2D rebinning is done. A check is added to
ensure this won't happen by accident. The Compton matrix is saved
instead of the complete response, for obvious reasons.

The Unfolder is rewritten to be Numba-friendly and more subclassable.
`Unfolder` is now an abstract base class, with `Guttormsen` being a
subclass implementing the traditional method.