Fix some bugs, refactor unfolding
There was a bug in how the components of the response matrix were
blended together. This is fixed.

The unfolding is being refactored to make it easier to both use and
implement new method. The old iteration method is implemented as well as
expection maximization.

Other minor bugs are also fixed.