Reorganize, rewrite first generation method.
Some files were moved around for organization.

The first generation method is rewritten, but the method is currently the same.
Is now a function returning a data class  instead of class instance. The valley
correction is removed, as the user should make the changes they want to their
matrix before first generation themselves.

Optional libraries are not imported by default. See `ompy.print_status()` for
a summary.

The `fileio` is an attempt for a unified writing/reading for all classes to and
from different formats. It turned out to be more finicky than I thought, so it
is not yet complete.

Minor changes to logging, removes some deprecated array functions, add
`to_root()` and `to_xarray()` for quicker interfacing with other code,
Guttormsen unfolding is slightly updated.