Add CACTUS response. Fix bugs.
The CACTUS response has been added, and can be loaded using
`om.response.Response.from_db('cactus')'`. The discrete
structures are interpolated using linear interpolation due
to the few data points. The low resolution of compton may
necessitate using other values for the lerping. This has not been tried.

LinearInterpolation and SplineInterpolation has been added.

FWHM should in the future be a composed Interpolation,
with an outer Scalable. CACTUS is currently the only response
using this feature.

The CPU version of compton interpolation is updated and working,
but has not been rigorously tested.