Rename names to make them agree with article
`Detector` is becoming the new interface to the user for creating the different matrices for unfolding. `Response` is being relegated to the creation and specialization of the discrete matrix. `Unfolder` needs to be revamped to treat `D`, `G_gamma` and `G_in` on equal footing.

Add Jax backend to `Detector.resolution_matrix`. In contrast to the numpy backend, it does a convolution with a Uniform distribution around the bin to account for the inherent midpoint uncertainty. Effect is completely negigble unless sigma ~ bin width, but even then the resulting distribution is so coarse that the difference is difficult to notice.