Several fixes
RMLE unfolding of 1D background was extended to be able to fold the
beta parameter.

`cut_like`, `reshape_like`, `align_with` are added to AbstractArray
as convenience functions to make arrays of different shapes congruent.

`pad=True` has been made default for Response, with a warning when the
extrapolation enters `e < 100 keV`, where artifacts begin to build.

Resampling of 1D spectra had a bug from the last refactor, where `nu`
was sampled instead of the total.